feat: add hazard and win overlay components
